+++ Assets/Scripts/Systems/Entities/EntityTarget.cs
@@ -7,6 +7,10 @@ namespace Metroidvania.Entities
         /// <summary>Cached transform of this object</summary>
         public Transform t { get; private set; }
 
+        [SerializeField] private Vector2 m_offset;
+
+        public Vector2 position => (Vector2)t.position + m_offset * t.localScale;
+
         private void Awake()
         {
             EntitiesManager.instance.AddTarget(this);
@@ -17,5 +21,14 @@ namespace Metroidvania.Entities
         {
             EntitiesManager.instance.RemoveTarget(this);
         }
+
+#if UNITY_EDITOR
+        private void OnDrawGizmosSelected()
+        {
+            new GizmosDrawer()
+                .SetColor(GizmosColor.instance.entities.targetPosition)
+                .DrawWireDisc((Vector2)transform.position + m_offset * transform.localScale, 0.1f);
+        }
+#endif
     }
 }