+++ Assets/Scripts/Systems/Entities/IEntityHittable.cs
@@ -1,0 +1,7 @@
+namespace Metroidvania.Entities
+{
+    public interface IEntityHittable
+    {
+        void OnTakeHit(EntityHitData hitData);
+    }
+}